pyproject.toml
[tool.poetry]
name = "sphiwe"
version = "0.1.0"
description = ""
authors = ["Nazire Mathe <nazireemathe@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-decouple = "^3.8"
django = "4.2.11"
django-split-settings = "^1.3.1"
dotmap = "^1.3.30"
pillow = "^10.3.0"
whitenoise = "^6.6.0"
requests = "^2.32.3"
django-crispy-forms = "^2.2"
crispy-bootstrap4 = "^2024.1"
channels = "^4.1.0"
moviepy = "^1.0.3"
google-auth = "^2.36.0"
qrcode = "^8.0"
django-otp = "^1.5.4"
google-cloud-storage = "^2.18.2"
django-axes = "^7.0.0"
boto = "^2.49.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
